converter-service/requirements.txt
fastapi
uvicorn
python-dotenv
boto3
python-docx
pdfminer.six
pymupdf
requests
